+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Make the log context map optional
 - Update Watermill logger to prepare for the next version
 - Export the Watermill logger type
+- Export the Invison logger type
 
 
 ## [0.8.0] - 2018-12-29