+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+
+## [0.10.0] - 2019-02-08
+
 ### Added
 
 - Separate interface for error logging
@@ -142,7 +145,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Initial release
 
 
-[Unreleased]: https://github.com/goph/logur/compare/v0.9.0...HEAD
+[Unreleased]: https://github.com/goph/logur/compare/v0.10.0...HEAD
+[0.10.0]: https://github.com/goph/logur/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/goph/logur/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/goph/logur/compare/v0.7.1...v0.8.0
 [0.7.1]: https://github.com/goph/logur/compare/v0.7.0...v0.7.1